source/getUsers/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2

environs==9.5.0
python-dotenv==1.0.0

loguru==0.7.2

SQLAlchemy==1.4.41
sqlmodel==0.0.8

# psycopg2==2.9.1
psycopg2-binary==2.9.1
asyncpg==0.29.0

Flask-Login==0.6.2
Werkzeug==2.3.7
